t718 Changed interval values to be doubles in java. Fixed bug where isDefined was returning true when max x is null. Fix issue where null template ticks was not allowing for save. Fixed old bug in condition block editor where adding an OR was adding an AND.

+++ eScreeningDashboard/src/main/java/gov/va/escreening/dto/template/GraphParamsDto.java
@@ -22,7 +22,7 @@ public class GraphParamsDto {
     private String footer;
     private Integer numberOfMonths;
     private String score;
-    private Map<String, Integer> intervals;
+    private Map<String, Double> intervals;
     private Double maxXPoint;
     private List<Double> ticks;
     
@@ -63,10 +63,10 @@ public class GraphParamsDto {
     public void setScore(String score) {
         this.score = score == null || score.startsWith("$") ? score : String.format(GRAPHICAL_FREEMARKER_WRAPPER, score);
     }
-    public Map<String, Integer> getIntervals() {
+    public Map<String, Double> getIntervals() {
         return intervals;
     }
-    public void setIntervals(Map<String, Integer> intervals) {
+    public void setIntervals(Map<String, Double> intervals) {
         this.intervals = intervals;
     }
     public Double getMaxXPoint() {
